+++ advanced-responsive-video-embedder.php
@@ -14,6 +14,7 @@
  * GitHub Branch:     master
  */
 
+
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
@@ -29,17 +30,17 @@ define( 'ARVE_ADMIN_URL',  ARVE_URL . 'admin/' );
 define( 'ARVE_NUM_TRACKS', 10 );
 
 if( ! class_exists( 'EDD_SL_Plugin_Updater' ) ) {
-	require_once plugin_dir_path( __FILE__ ) . 'admin/class-edd-sl-plugin-updater.php';
+	require_once __DIR__ . '/admin/class-edd-sl-plugin-updater.php';
 }
 
+#add_action( 'init', 'arve_init' );
 arve_init();
-#add_action( 'init', 'arve_init' ); # TODO ??
 
 function arve_init() {
 
 	add_option( 'arve_install_date', current_time( 'timestamp' ) );
 
-	require_once plugin_dir_path( __FILE__ ) . 'admin/class-arve-admin-notice-factory.php';
+	require_once __DIR__ . '/admin/class-arve-admin-notice-factory.php';
 
 	if ( version_compare( PHP_VERSION, '5.3', '<' ) ) {
 		add_action( 'admin_init', 'arve_php_outdated_message' );