+++ public/functions-shortcodes.php
@@ -24,15 +24,16 @@ function arve_shortcode_arve( $input_atts, $content = null, $arve_shortcode = tr
 		'title'         => null,
 		'upload_date'   => null,
 		# <video>
-		'm4v'         => null,
-		'mp4'         => null,
-		'ogv'         => null,
-		'webm'        => null,
-		'preload'     => 'metadata',
-		'playsinline' => null,
-		'muted'       => null,
-		'controls'    => 'y',
-		'loop'        => 'n',
+		'm4v'          => null,
+		'mp4'          => null,
+		'ogv'          => null,
+		'webm'         => null,
+		'preload'      => 'metadata',
+		'playsinline'  => null,
+		'muted'        => null,
+		'controls'     => 'y',
+		'controlslist' => empty( $options['controlslist'] ) ? null : (string) $options['controlslist'],
+		'loop'         => 'n',
 		# TED only
 		'lang'     => null,
 		# Vimeo only
@@ -151,7 +152,7 @@ function arve_shortcode_arve_supported() {
 	  $out .= '<tr>';
 	  $out .= sprintf( '<td>%d</td>', $count++ );
 	  $out .= sprintf( '<td>%s</td>', esc_html( $values['name'] ) );
-	  $out .= sprintf( '<td>%s</td>', ( isset( $values['requires_src'] ) && $values['requires_src'] ) ? '' : '&#x2713;' );
+	  $out .= sprintf( '<td>%s</td>', ( isset( $values['requires_src'] ) && $values['requires_src'] ) ? '&#x2713;' : '' );
 	  $out .= sprintf( '<td>%s</td>', ( isset( $values['embed_url'] ) && arve_starts_with( $values['embed_url'], 'https' ) ) ? '&#x2713;' : '' );
 	  $out .= sprintf( '<td>%s</td>', ! empty( $values['requires_flash'] ) ? '&#x2713;' : '' );
 	  $out .= sprintf( '<td>%s</td>', ( isset( $values['auto_thumbnail'] ) && $values['auto_thumbnail'] ) ? '&#x2713;' : '' );