+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### 2017-05-05 - 8.5.2 ###
 
-* Improved: Embed Liveleak with https.
+* Improved: Embed Liveleak with https. Remove the default parameter.
 
 ### 2017-05-05 - 8.5.1 ###